+++ packages/core/tooling/parsers.ts
@@ -1,4 +1,4 @@
-import * as tools from './tools.ts';
+import * as utils from './utils.ts';
 import MagicString from 'magic-string';
 
 type ParseBase = {
@@ -6,31 +6,31 @@ type ParseBase = {
 	generateCode(): string;
 };
 
-export function parseScript(source: string): { ast: tools.AstTypes.Program } & ParseBase {
-	const ast = tools.parseScript(source);
-	const generateCode = () => tools.serializeScript(ast, source);
+export function parseScript(source: string): { ast: utils.AstTypes.Program } & ParseBase {
+	const ast = utils.parseScript(source);
+	const generateCode = () => utils.serializeScript(ast, source);
 
 	return { ast, source, generateCode };
 }
 
-export function parseCss(source: string): { ast: tools.CssAst } & ParseBase {
-	const ast = tools.parseCss(source);
+export function parseCss(source: string): { ast: utils.CssAst } & ParseBase {
+	const ast = utils.parseCss(source);
 	const generateCode = () => ast.toString();
 
 	return { ast, source, generateCode };
 }
 
-export function parseHtml(source: string): { ast: tools.HtmlDocument } & ParseBase {
-	const ast = tools.parseHtml(source);
-	const generateCode = () => tools.serializeHtml(ast);
+export function parseHtml(source: string): { ast: utils.HtmlDocument } & ParseBase {
+	const ast = utils.parseHtml(source);
+	const generateCode = () => utils.serializeHtml(ast);
 
 	return { ast, source, generateCode };
 }
 
 export function parseJson(source: string): { data: any } & ParseBase {
 	if (!source) source = '{}';
-	const data = tools.parseJson(source);
-	const generateCode = () => tools.serializeJson(source, data);
+	const data = utils.parseJson(source);
+	const generateCode = () => utils.serializeJson(source, data);
 
 	return { data, source, generateCode };
 }
